+++ CHANGELOG.md
@@ -162,6 +162,7 @@
 - Fix Players disconnecting from servers if they made changes to certain config options
 - Fix JSON stores not loading properly if player with shouldered Pokémon is migrating from 1.5.2 into 1.6
 - Fix PokemonProperties ability=myAbility being declared as forced despite being legal for the Pokémon
+- Fixed type formatting in Pokédex scanner mode when dual types require two lines
 
 ### Developer
 - Updated the Pokédex data updated events to always include a Pokemon instance, and optionally a DisguiseData instance.